requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
certifi==2020.12.5
chardet==3.0.4
deep-translator==1.3.2
discord.py==1.6.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
MouseInfo==0.1.3
multidict==5.1.0
numexpr3==3.0.1a6
numpy==1.19.5
Pillow==8.1.0
praw==7.1.0
prawcore==1.5.0
PyAutoGUI==0.9.52
PyDirectInput==1.0.3
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pywin32==300
requests==2.25.1
rfc3986==1.4.0
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
stemming==1.0.1
typing-extensions==3.7.4.3
update-checker==0.18.0
urllib3==1.26.2
websocket-client==0.57.0
yarl==1.6.3
